Various fixes #2
