+++ app/src/main/java/android/academy/nytimes/mvp/Model.kt
@@ -1,13 +1,13 @@
 package android.academy.nytimes.mvp
 
 import android.academy.nytimes.data.NewsItem
-import android.academy.nytimes.utils.Constants
+import android.academy.nytimes.utils.Category
 import io.reactivex.Single
 import javax.inject.Inject
 
 class Model @Inject constructor(private val remoteRepository: RemoteRepository) {
 
-    fun loadNews(category: Constants.Category): Single<List<NewsItem>> {
+    fun loadNews(category: Category): Single<List<NewsItem>> {
         return remoteRepository.getNews(category)
     }
