+++ app/src/main/java/android/academy/nytimes/network/NyTimesInterceptor.kt
@@ -1,11 +1,12 @@
 package android.academy.nytimes.network
 
+import android.academy.nytimes.utils.Constants
 import okhttp3.Interceptor
 import okhttp3.Response
 import java.io.IOException
 import javax.inject.Inject
 
-class NyTimesInterceptor @Inject constructor(private val token: String)
+class NyTimesInterceptor @Inject constructor()
     : Interceptor {
 
     companion object {
@@ -15,7 +16,7 @@ class NyTimesInterceptor @Inject constructor(private val token: String)
     @Throws(IOException::class)
     override fun intercept(chain: Interceptor.Chain): Response {
         val original = chain.request()
-        val request = original.newBuilder().addHeader(REQUEST_HEADER, token).build()
+        val request = original.newBuilder().addHeader(REQUEST_HEADER, Constants.TOKEN).build()
         return chain.proceed(request)
     }
 }
