+++ app/src/main/java/android/academy/nytimes/utils/ContextExt.kt
@@ -1,0 +1,7 @@
+package android.academy.nytimes.utils
+
+import android.content.Context
+import android.content.res.Configuration
+
+fun Context.isHorizontalOrientation() =
+        resources.configuration.orientation == Configuration.ORIENTATION_LANDSCAPE
