+++ app/src/main/java/android/academy/nytimes/utils/StringExt.kt
@@ -1,0 +1,3 @@
+package android.academy.nytimes.utils
+
+fun String.Companion.empty() = ""
